Fixed form for small screens
The iframe of the form had a minimum width.
This caused the page to break on devices with smaller screens.

To remove the scrollbar in the iframe, I had to set the height of the iframe to a greater, fixed value.
Unfortunately, the height of the iframe can't be dynamically decided based on the content of the iframe, because of "Cross domain policy".